callout: add a new module which is an alternative to the scheduler
The 'callout' module is another timer manager that aims to replace the
scheduler module. The management of timers can be done either from an
asynchronous interruption or from the main program. This module is
able to handle priorities between timers and allow to reschedule
a timer later if its priority is too low.

Signed-off-by: Olivier Matz <zer0@droids-corp.org>